adding point cloud visualizer

+++ include/altimeter_processing.h
@@ -9,6 +9,6 @@ class AltimeterProcessing{
 public:
     AltimeterProcessing();
     ~AltimeterProcessing();
-    std::vector<double> depth;
+    std::vector<float> depth;
     void load_altimeter(std::string &filedir, int & data_num);
 };
